debugger: Test GetHandlers()
